requirements.txt
absl-py==0.13.0
aio-pika==6.8.2
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiormq==3.3.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.7.0
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
bidict==0.23.1
boto3==1.38.24
botocore==1.38.24
build==1.2.2.post1
CacheControl==0.12.14
cachetools==4.2.4
certifi==2025.4.26
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
chromadb==1.0.11
click==8.1.8
cloudpickle==1.6.0
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.0.4
cryptography==45.0.3
cycler==0.12.1
dask==2022.2.0
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
dnspython==1.16.0
docopt==0.6.2
durationpy==0.10
exceptiongroup==1.3.0
fastapi==0.115.9
fbmessenger==6.0.0
filelock==3.18.0
fire==0.7.0
flatbuffers==2.0.7
frozenlist==1.6.0
fsspec==2025.5.1
future==1.0.0
fuzzywuzzy==0.18.0
gast==0.4.0
google-auth==1.35.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.2
groq==0.25.0
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
hf-xet==1.1.2
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.0.1
jsonpatch==1.33
jsonpickle==2.0.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kafka-python==2.2.10
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
kubernetes==32.0.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.62
langchain-text-splitters==0.3.8
langsmith==0.3.42
libclang==18.1.1
locket==1.0.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.3.4
mattermostwrapper==2.2
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
msgpack==1.1.0
multidict==5.2.0
mypy_extensions==1.1.0
networkx==2.6.3
numpy==1.23.5
oauthlib==3.2.2
onnxruntime==1.19.2
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
opt_einsum==3.4.0
orjson==3.10.18
overrides==7.7.0
packaging==21.3
pamqp==2.3.0
partd==1.4.2
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
pluggy==1.6.0
posthog==4.2.0
prompt-toolkit==2.0.10
propcache==0.3.1
protobuf==3.20.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydot==1.4.2
Pygments==2.19.1
PyJWT==2.10.1
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.2.3
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyrsistent==0.20.0
pyTelegramBotAPI==4.27.0
pytest==8.4.0
python-crfsuite==0.9.11
python-dateutil==2.8.2
python-dotenv==1.1.0
python-engineio==4.12.1
python-socketio==5.13.0
pytz==2021.3
PyYAML==6.0.2
questionary==1.10.0
randomname==0.1.5
rasa==3.2.0
rasa-sdk==3.2.3
redis==3.5.3
referencing==0.36.2
regex==2022.4.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rocketchat-API==1.25.0
rpds-py==0.25.1
rsa==4.9.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.12
s3transfer==0.13.0
safetensors==0.5.3
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==0.24.2
scipy==1.7.3
sentry-sdk==1.3.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sklearn-crfsuite==0.3.6
slackclient==2.9.4
sniffio==1.3.1
SQLAlchemy==1.4.54
starlette==0.45.3
sympy==1.14.0
tabulate==0.9.0
tarsafe==0.0.3
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.7.4
tensorflow-addons==0.15.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-text==2.7.3
termcolor==3.1.0
terminaltables==3.1.10
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
toolz==1.0.0
tqdm==4.67.1
transformers==4.52.4
twilio==6.50.1
typeguard==2.13.3
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing-utils==0.1.0
typing_extensions==4.13.2
tzlocal==2.1
ujson==5.10.0
urllib3==1.26.20
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
webexteamssdk==1.6.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
